+++ src/Resource/ResourceCompiler.php
@@ -89,7 +89,6 @@ final class ResourceCompiler
             'inlineUpdate' => ['method' => ['post'], 'middleware' => [$update]],
             'replicate' => ['method' => ['post'], 'middleware' => [$replicate]],
             'reorder' => ['method' => ['post'], 'middleware' => [$reorder]],
-            'exportCsv' => ['method' => ['get', 'post'], 'middleware' => [$view]],
             'export' => ['method' => ['get', 'post'], 'middleware' => [$view]],
             'delete' => ['method' => ['post'], 'middleware' => [$delete]],
             'restore' => ['method' => ['post'], 'middleware' => [$restore]],